a start for #11 and add rectangles around glyphs This commit is an attempt to add user interface for configuration of some parameters such as rule colors, widths and off/on logic per node type.
It is using the luakeval module, a module which
is format agnostic, so it should work in plain
and LaTeX.

Some parameters are still not implemented.
Mostly options for kern, glue and penalty
should be refined, as they have colors for
different situations, so more keys needs to
be added.

In addition documentation should be updated.